requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.4
pkg_resources==0.0.0
PyJWT==2.5.0
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.0.1
